start separation of event rendered view and grid view
