+++ x-pack/plugins/timelines/public/components/t_grid/body/index.tsx
@@ -5,90 +5,33 @@
  * 2.0.
  */
 
-import {
-  EuiDataGridRefProps,
-  EuiDataGridColumn,
-  EuiDataGridCellValueElementProps,
-  EuiDataGridControlColumn,
-  EuiDataGridStyle,
-  EuiDataGridToolBarVisibilityOptions,
-  EuiLoadingSpinner,
-  EuiFlexGroup,
-  EuiFlexItem,
-  EuiProgress,
-  EuiDataGridSorting,
-} from '@elastic/eui';
-import { getOr } from 'lodash/fp';
-import memoizeOne from 'memoize-one';
-import React, {
-  ComponentType,
-  lazy,
-  Suspense,
-  useCallback,
-  useEffect,
-  useMemo,
-  useState,
-  useContext,
-  useRef,
-} from 'react';
-import type { ReactNode } from 'react';
-import { connect, ConnectedProps, useDispatch } from 'react-redux';
-
-import styled, { ThemeContext } from 'styled-components';
-import { ALERT_RULE_CONSUMER, ALERT_RULE_PRODUCER } from '@kbn/rule-data-utils';
+import React, { lazy } from 'react';
 import { Filter } from '@kbn/es-query';
-import type { EuiTheme } from '@kbn/kibana-react-plugin/common';
-import { FieldBrowserOptions, FetchAlertData } from '@kbn/triggers-actions-ui-plugin/public';
-import { useKibana } from '@kbn/kibana-react-plugin/public';
-import { EcsFieldsResponse } from '@kbn/rule-registry-plugin/common/search_strategy';
+import { FieldBrowserOptions } from '@kbn/triggers-actions-ui-plugin/public';
+
 import {
   TGridCellAction,
   BulkActionsProp,
   CellValueElementProps,
-  ColumnHeaderOptions,
   ControlColumnProps,
   RowRenderer,
   AlertStatus,
-  SortColumnTimeline,
-  TimelineId,
   TimelineTabs,
-  SetEventsLoading,
-  SetEventsDeleted,
 } from '../../../../common/types/timeline';
-
-import type { TimelineItem, TimelineNonEcsData } from '../../../../common/search_strategy/timeline';
-
-import { getColumnHeader, getColumnHeaders } from './column_headers/helpers';
-import {
-  addBuildingBlockStyle,
-  getEventIdToDataMapping,
-  hasCellActions,
-  mapSortDirectionToDirection,
-  mapSortingColumns,
-} from './helpers';
-
+import type { TimelineItem } from '../../../../common/search_strategy/timeline';
 import type { BrowserFields } from '../../../../common/search_strategy/index_fields';
-import type { OnRowSelected, OnSelectAll } from '../types';
 import type { Refetch } from '../../../store/t_grid/inputs';
 import { Ecs } from '../../../../common/ecs';
-import { getPageRowIndex } from '../../../../common/utils/pagination';
-import { StatefulEventContext } from '../../stateful_event_context';
-import { tGridActions, TGridModel, tGridSelectors, TimelineState } from '../../../store/t_grid';
-import { useDeepEqualSelector } from '../../../hooks/use_selector';
-import { RowAction } from './row_action';
-import * as i18n from './translations';
-import { AlertCount } from '../styles';
-import { checkBoxControlColumn } from './control_columns';
 import { ViewSelection } from '../event_rendered_view/selector';
-import { EventRenderedView } from '../event_rendered_view';
-import { REMOVE_COLUMN } from './column_headers/translations';
-import { TimelinesStartPlugins } from '../../../types';
-import { TGridComponentStateProvider } from '../../../methods/context';
-import { ALERT_TABLE_CONFIGURATION_KEY } from '../config';
 
-const StatefulAlertBulkActions = lazy(() => import('../toolbar/bulk_actions/alert_bulk_actions'));
+const StatefulGridBody = lazy(async () => import('./grid'));
+const StatefuEventRenderedBody = lazy(async () => import('./event_rendered'));
 
-interface OwnProps {
+// NEXT UP:
+// - Move fetching into Grid / Event Rendered (one uses useTimeline blah, one should only use the generated query)
+// - remove unused props
+
+export interface StatefulBodyProps {
   activePage: number;
   additionalControls?: React.ReactNode;
   appId?: string;
@@ -135,868 +78,22 @@ interface OwnProps {
   showCheckboxes?: boolean;
 }
 
-const defaultUnit = (n: number) => i18n.ALERTS_UNIT(n);
-
-export const hasAdditionalActions = (id: TimelineId): boolean =>
-  [TimelineId.detectionsPage, TimelineId.detectionsRulesDetailsPage, TimelineId.active].includes(
-    id
-  );
-
-const ES_LIMIT_COUNT = 9999;
-
-const EMPTY_CONTROL_COLUMNS: ControlColumnProps[] = [];
-
-const EmptyHeaderCellRender: ComponentType = () => null;
-
-const gridStyle: EuiDataGridStyle = { border: 'none', fontSize: 's', header: 'underline' };
-
-const EuiDataGridContainer = styled.div<{ hideLastPage: boolean }>`
-  ul.euiPagination__list {
-    li.euiPagination__item:last-child {
-      ${({ hideLastPage }) => `${hideLastPage ? 'display:none' : ''}`};
-    }
-  }
-`;
-
-const transformControlColumns = ({
-  columnHeaders,
-  controlColumns,
-  data,
-  fieldBrowserOptions,
-  isEventViewer = false,
-  loadingEventIds,
-  onRowSelected,
-  onRuleChange,
-  selectedEventIds,
-  showCheckboxes,
-  tabType,
-  timelineId,
-  isSelectAllChecked,
-  onSelectPage,
-  browserFields,
-  pageSize,
-  sort,
-  theme,
-  setEventsLoading,
-  setEventsDeleted,
-  hasAlertsCrudPermissions,
-}: {
-  columnHeaders: ColumnHeaderOptions[];
-  controlColumns: ControlColumnProps[];
-  data: TimelineItem[];
-  disabledCellActions: string[];
-  fieldBrowserOptions?: FieldBrowserOptions;
-  isEventViewer?: boolean;
-  loadingEventIds: string[];
-  onRowSelected: OnRowSelected;
-  onRuleChange?: () => void;
-  selectedEventIds: Record<string, TimelineNonEcsData[]>;
-  showCheckboxes: boolean;
-  tabType: TimelineTabs;
-  timelineId: string;
-  isSelectAllChecked: boolean;
-  browserFields: BrowserFields;
-  onSelectPage: OnSelectAll;
-  pageSize: number;
-  sort: SortColumnTimeline[];
-  theme: EuiTheme;
-  setEventsLoading: SetEventsLoading;
-  setEventsDeleted: SetEventsDeleted;
-  hasAlertsCrudPermissions?: ({
-    ruleConsumer,
-    ruleProducer,
-  }: {
-    ruleConsumer: string;
-    ruleProducer?: string;
-  }) => boolean;
-}): EuiDataGridControlColumn[] =>
-  controlColumns.map(
-    ({ id: columnId, headerCellRender = EmptyHeaderCellRender, rowCellRender, width }, i) => ({
-      id: `${columnId}`,
-      headerCellRender: () => {
-        const HeaderActions = headerCellRender;
-        return (
-          <>
-            {HeaderActions && (
-              <HeaderActions
-                width={width}
-                browserFields={browserFields}
-                fieldBrowserOptions={fieldBrowserOptions}
-                columnHeaders={columnHeaders}
-                isEventViewer={isEventViewer}
-                isSelectAllChecked={isSelectAllChecked}
-                onSelectAll={onSelectPage}
-                showEventsSelect={false}
-                showSelectAllCheckbox={showCheckboxes}
-                sort={sort}
-                tabType={tabType}
-                timelineId={timelineId}
-              />
-            )}
-          </>
-        );
-      },
-      rowCellRender: ({
-        isDetails,
-        isExpandable,
-        isExpanded,
-        rowIndex,
-        colIndex,
-        setCellProps,
-      }: EuiDataGridCellValueElementProps) => {
-        const pageRowIndex = getPageRowIndex(rowIndex, pageSize);
-        const rowData = data[pageRowIndex];
-
-        let disabled = false;
-        if (rowData) {
-          addBuildingBlockStyle(rowData.ecs, theme, setCellProps);
-          if (columnId === 'checkbox-control-column' && hasAlertsCrudPermissions != null) {
-            // FUTURE ENGINEER, the assumption here is you can only have one producer and consumer at this time
-            const ruleConsumers =
-              rowData.data.find((d) => d.field === ALERT_RULE_CONSUMER)?.value ?? [];
-            const ruleProducers =
-              rowData.data.find((d) => d.field === ALERT_RULE_PRODUCER)?.value ?? [];
-            disabled = !hasAlertsCrudPermissions({
-              ruleConsumer: ruleConsumers.length > 0 ? ruleConsumers[0] : '',
-              ruleProducer: ruleProducers.length > 0 ? ruleProducers[0] : undefined,
-            });
-          }
-        } else {
-          // disable the cell when it has no data
-          setCellProps({ style: { display: 'none' } });
-        }
-
-        return (
-          <RowAction
-            columnId={columnId ?? ''}
-            columnHeaders={columnHeaders}
-            controlColumn={controlColumns[i]}
-            data={data}
-            disabled={disabled}
-            index={i}
-            isDetails={isDetails}
-            isExpanded={isExpanded}
-            isEventViewer={isEventViewer}
-            isExpandable={isExpandable}
-            loadingEventIds={loadingEventIds}
-            onRowSelected={onRowSelected}
-            onRuleChange={onRuleChange}
-            rowIndex={rowIndex}
-            colIndex={colIndex}
-            pageRowIndex={pageRowIndex}
-            selectedEventIds={selectedEventIds}
-            setCellProps={setCellProps}
-            showCheckboxes={showCheckboxes}
-            tabType={tabType}
-            timelineId={timelineId}
-            width={width}
-            setEventsLoading={setEventsLoading}
-            setEventsDeleted={setEventsDeleted}
-          />
-        );
-      },
-      width,
-    })
-  );
-
-export type StatefulBodyProps = OwnProps & PropsFromRedux;
-
 /**
  * The Body component is used everywhere timeline is used within the security application. It is the highest level component
  * that is shared across all implementations of the timeline.
  */
 
 export const BodyComponent = React.memo<StatefulBodyProps>(
-  ({
-    activePage,
-    additionalControls,
-    appId = '',
-    browserFields,
-    bulkActions = true,
-    clearSelected,
-    columnHeaders,
-    data,
-    defaultCellActions,
-    disabledCellActions,
-    fieldBrowserOptions,
-    filterQuery,
-    filters,
-    filterStatus,
-    getRowRenderer,
-    hasAlertsCrud,
-    hasAlertsCrudPermissions,
-    id,
-    indexNames,
-    isEventViewer = false,
-    isLoading,
-    isSelectAllChecked,
-    itemsPerPageOptions,
-    leadingControlColumns = EMPTY_CONTROL_COLUMNS,
-    loadingEventIds,
-    loadPage,
-    onRuleChange,
-    pageSize,
-    refetch,
-    renderCellValue,
-    rowRenderers,
-    selectedEventIds,
-    setSelected,
-    showCheckboxes,
-    sort,
-    tableView = 'gridView',
-    tabType,
-    totalItems,
-    totalSelectAllAlerts,
-    unit = defaultUnit,
-  }) => {
-    const { triggersActionsUi } = useKibana<TimelinesStartPlugins>().services;
-
-    const dataGridRef = useRef<EuiDataGridRefProps>(null);
-
-    const dispatch = useDispatch();
-    const getManageTimeline = useMemo(() => tGridSelectors.getManageTimelineById(), []);
-    const { queryFields, selectAll, defaultColumns } = useDeepEqualSelector((state) =>
-      getManageTimeline(state, id)
-    );
-
-    const alertCountText = useMemo(
-      () => `${totalItems.toLocaleString()} ${unit(totalItems)}`,
-      [totalItems, unit]
-    );
-
-    const selectedCount = useMemo(() => Object.keys(selectedEventIds).length, [selectedEventIds]);
-
-    const theme: EuiTheme = useContext(ThemeContext);
-    const onRowSelected: OnRowSelected = useCallback(
-      ({ eventIds, isSelected }: { eventIds: string[]; isSelected: boolean }) => {
-        setSelected({
-          id,
-          eventIds: getEventIdToDataMapping(
-            data,
-            eventIds,
-            queryFields,
-            hasAlertsCrud ?? false,
-            hasAlertsCrudPermissions
-          ),
-          isSelected,
-          isSelectAllChecked: isSelected && selectedCount + 1 === data.length,
-        });
-      },
-      [setSelected, id, data, queryFields, hasAlertsCrud, hasAlertsCrudPermissions, selectedCount]
-    );
-
-    const onSelectPage: OnSelectAll = useCallback(
-      ({ isSelected }: { isSelected: boolean }) =>
-        isSelected
-          ? setSelected({
-              id,
-              eventIds: getEventIdToDataMapping(
-                data,
-                data.map((event) => event._id),
-                queryFields,
-                hasAlertsCrud ?? false,
-                hasAlertsCrudPermissions
-              ),
-              isSelected,
-              isSelectAllChecked: isSelected,
-            })
-          : clearSelected({ id }),
-      [setSelected, id, data, queryFields, hasAlertsCrud, hasAlertsCrudPermissions, clearSelected]
-    );
-
-    // Sync to selectAll so parent components can select all events
-    useEffect(() => {
-      if (selectAll && !isSelectAllChecked) {
-        onSelectPage({ isSelected: true });
-      }
-    }, [isSelectAllChecked, onSelectPage, selectAll]);
-
-    const onAlertStatusActionSuccess = useMemo(() => {
-      if (bulkActions && bulkActions !== true) {
-        return bulkActions.onAlertStatusActionSuccess;
-      }
-    }, [bulkActions]);
-
-    const onAlertStatusActionFailure = useMemo(() => {
-      if (bulkActions && bulkActions !== true) {
-        return bulkActions.onAlertStatusActionFailure;
-      }
-    }, [bulkActions]);
-
-    const additionalBulkActions = useMemo(() => {
-      if (bulkActions && bulkActions !== true && bulkActions.customBulkActions !== undefined) {
-        return bulkActions.customBulkActions.map((action) => {
-          return {
-            ...action,
-            onClick: (eventIds: string[]) => {
-              const items = data.filter((item) => {
-                return eventIds.find((event) => item._id === event);
-              });
-              action.onClick(items);
-            },
-          };
-        });
-      }
-    }, [bulkActions, data]);
-
-    const showAlertStatusActions = useMemo(() => {
-      if (!hasAlertsCrud) {
-        return false;
-      }
-      if (typeof bulkActions === 'boolean') {
-        return bulkActions;
-      }
-      return bulkActions.alertStatusActions ?? true;
-    }, [bulkActions, hasAlertsCrud]);
-
-    const showBulkActions = useMemo(() => {
-      if (!hasAlertsCrud) {
-        return false;
-      }
-
-      if (selectedCount === 0 || !showCheckboxes) {
-        return false;
-      }
-      if (typeof bulkActions === 'boolean') {
-        return bulkActions;
-      }
-      return (bulkActions?.customBulkActions?.length || bulkActions?.alertStatusActions) ?? true;
-    }, [hasAlertsCrud, selectedCount, showCheckboxes, bulkActions]);
-
-    const onResetColumns = useCallback(() => {
-      dispatch(tGridActions.updateColumns({ id, columns: defaultColumns }));
-    }, [defaultColumns, dispatch, id]);
-
-    const onToggleColumn = useCallback(
-      (columnId: string) => {
-        if (columnHeaders.some(({ id: columnHeaderId }) => columnId === columnHeaderId)) {
-          dispatch(
-            tGridActions.removeColumn({
-              columnId,
-              id,
-            })
-          );
-        } else {
-          dispatch(
-            tGridActions.upsertColumn({
-              column: getColumnHeader(columnId, defaultColumns),
-              id,
-              index: 1,
-            })
-          );
-        }
-      },
-      [columnHeaders, dispatch, id, defaultColumns]
-    );
-
-    const alertToolbar = useMemo(
-      () => (
-        <EuiFlexGroup gutterSize="m" alignItems="center">
-          <EuiFlexItem grow={false}>
-            <AlertCount>{alertCountText}</AlertCount>
-          </EuiFlexItem>
-          {showBulkActions && (
-            <Suspense fallback={<EuiLoadingSpinner />}>
-              <StatefulAlertBulkActions
-                showAlertStatusActions={showAlertStatusActions}
-                data-test-subj="bulk-actions"
-                id={id}
-                totalItems={totalSelectAllAlerts ?? totalItems}
-                filterStatus={filterStatus}
-                query={filterQuery}
-                indexName={indexNames.join()}
-                onActionSuccess={onAlertStatusActionSuccess}
-                onActionFailure={onAlertStatusActionFailure}
-                customBulkActions={additionalBulkActions}
-                refetch={refetch}
-              />
-            </Suspense>
-          )}
-        </EuiFlexGroup>
-      ),
-      [
-        additionalBulkActions,
-        alertCountText,
-        filterQuery,
-        filterStatus,
-        id,
-        indexNames,
-        onAlertStatusActionFailure,
-        onAlertStatusActionSuccess,
-        refetch,
-        showAlertStatusActions,
-        showBulkActions,
-        totalItems,
-        totalSelectAllAlerts,
-      ]
-    );
-
-    const toolbarVisibility: EuiDataGridToolBarVisibilityOptions = useMemo(
-      () => ({
-        additionalControls: (
-          <>
-            {isLoading && <EuiProgress size="xs" position="absolute" color="accent" />}
-            <AlertCount data-test-subj="server-side-event-count">{alertCountText}</AlertCount>
-            {showBulkActions ? (
-              <>
-                <Suspense fallback={<EuiLoadingSpinner />}>
-                  <StatefulAlertBulkActions
-                    showAlertStatusActions={showAlertStatusActions}
-                    data-test-subj="bulk-actions"
-                    id={id}
-                    totalItems={totalSelectAllAlerts ?? totalItems}
-                    filterStatus={filterStatus}
-                    query={filterQuery}
-                    indexName={indexNames.join()}
-                    onActionSuccess={onAlertStatusActionSuccess}
-                    onActionFailure={onAlertStatusActionFailure}
-                    customBulkActions={additionalBulkActions}
-                    refetch={refetch}
-                  />
-                </Suspense>
-                {additionalControls ?? null}
-              </>
-            ) : (
-              <>
-                {additionalControls ?? null}
-                {triggersActionsUi.getFieldBrowser({
-                  browserFields,
-                  options: fieldBrowserOptions,
-                  columnIds: columnHeaders.map(({ id: columnId }) => columnId),
-                  onResetColumns,
-                  onToggleColumn,
-                })}
-              </>
-            )}
-          </>
-        ),
-        ...(showBulkActions
-          ? {
-              showColumnSelector: false,
-              showSortSelector: false,
-              showFullScreenSelector: false,
-            }
-          : {
-              showColumnSelector: { allowHide: false, allowReorder: true },
-              showSortSelector: true,
-              showFullScreenSelector: true,
-            }),
-        showDisplaySelector: false,
-      }),
-      [
-        isLoading,
-        alertCountText,
-        showBulkActions,
-        showAlertStatusActions,
-        id,
-        totalSelectAllAlerts,
-        totalItems,
-        filterStatus,
-        filterQuery,
-        indexNames,
-        onAlertStatusActionSuccess,
-        onAlertStatusActionFailure,
-        onResetColumns,
-        onToggleColumn,
-        triggersActionsUi,
-        additionalBulkActions,
-        refetch,
-        additionalControls,
-        browserFields,
-        fieldBrowserOptions,
-        columnHeaders,
-      ]
-    );
-
-    const sortingColumns: Array<{
-      id: string;
-      direction: 'asc' | 'desc';
-    }> = useMemo(
-      () =>
-        sort.map((x) => ({
-          id: x.columnId,
-          direction: mapSortDirectionToDirection(x.sortDirection),
-        })),
-      [sort]
-    );
-
-    const onSort = useCallback(
-      (
-        nextSortingColumns: Array<{
-          id: string;
-          direction: 'asc' | 'desc';
-        }>
-      ) => {
-        dispatch(
-          tGridActions.updateSort({
-            id,
-            sort: mapSortingColumns({ columns: nextSortingColumns, columnHeaders }),
-          })
-        );
-
-        setTimeout(() => {
-          // schedule the query to be re-executed from page 0, (but only after the
-          // store has been updated with the new sort):
-          if (loadPage != null) {
-            loadPage(0);
-          }
-        }, 0);
-      },
-      [columnHeaders, dispatch, id, loadPage]
-    );
-
-    const visibleColumns = useMemo(() => columnHeaders.map(({ id: cid }) => cid), [columnHeaders]); // the full set of columns
-
-    const onColumnResize = useCallback(
-      ({ columnId, width }: { columnId: string; width: number }) => {
-        dispatch(
-          tGridActions.updateColumnWidth({
-            columnId,
-            id,
-            width,
-          })
-        );
-      },
-      [dispatch, id]
-    );
-
-    const onSetVisibleColumns = useCallback(
-      (newVisibleColumns: string[]) => {
-        dispatch(
-          tGridActions.updateColumnOrder({
-            columnIds: newVisibleColumns,
-            id,
-          })
-        );
-      },
-      [dispatch, id]
-    );
-
-    const setEventsLoading = useCallback<SetEventsLoading>(
-      ({ eventIds, isLoading: loading }) => {
-        dispatch(tGridActions.setEventsLoading({ id, eventIds, isLoading: loading }));
-      },
-      [dispatch, id]
-    );
-
-    const setEventsDeleted = useCallback<SetEventsDeleted>(
-      ({ eventIds, isDeleted }) => {
-        dispatch(tGridActions.setEventsDeleted({ id, eventIds, isDeleted }));
-      },
-      [dispatch, id]
-    );
-
-    const leadingTGridControlColumns = useMemo(() => {
-      const controlColumns = showCheckboxes
-        ? [checkBoxControlColumn, ...leadingControlColumns]
-        : leadingControlColumns;
-      return transformControlColumns({
-        columnHeaders,
-        controlColumns,
-        data,
-        disabledCellActions,
-        fieldBrowserOptions,
-        isEventViewer,
-        loadingEventIds,
-        onRowSelected,
-        onRuleChange,
-        selectedEventIds,
-        showCheckboxes,
-        tabType,
-        timelineId: id,
-        isSelectAllChecked,
-        sort,
-        browserFields,
-        onSelectPage,
-        theme,
-        setEventsLoading,
-        setEventsDeleted,
-        pageSize,
-        hasAlertsCrudPermissions,
-      });
-    }, [
-      showCheckboxes,
-      leadingControlColumns,
-      columnHeaders,
-      data,
-      disabledCellActions,
-      fieldBrowserOptions,
-      isEventViewer,
-      id,
-      loadingEventIds,
-      onRowSelected,
-      onRuleChange,
-      selectedEventIds,
-      tabType,
-      isSelectAllChecked,
-      sort,
-      browserFields,
-      onSelectPage,
-      theme,
-      pageSize,
-      setEventsLoading,
-      setEventsDeleted,
-      hasAlertsCrudPermissions,
-    ]);
-    const columnsWithCellActions: EuiDataGridColumn[] = useMemo(
-      () =>
-        columnHeaders.map((header) => {
-          const buildAction = (tGridCellAction: TGridCellAction) =>
-            tGridCellAction({
-              browserFields,
-              data: data.map((row) => row.data),
-              ecsData: data.map((row) => row.ecs),
-              header: columnHeaders.find((h) => h.id === header.id),
-              pageSize,
-              timelineId: id,
-              closeCellPopover: dataGridRef.current?.closeCellPopover,
-            });
-          return {
-            ...header,
-            actions: {
-              ...header.actions,
-              additional: [
-                {
-                  iconType: 'cross',
-                  label: REMOVE_COLUMN,
-                  onClick: () => {
-                    dispatch(tGridActions.removeColumn({ id, columnId: header.id }));
-                  },
-                  size: 'xs',
-                },
-              ],
-            },
-            ...(hasCellActions({
-              columnId: header.id,
-              disabledCellActions,
-            })
-              ? {
-                  cellActions:
-                    header.tGridCellActions?.map(buildAction) ??
-                    defaultCellActions?.map(buildAction),
-                  visibleCellActions: 3,
-                }
-              : {}),
-          };
-        }),
-      [
-        browserFields,
-        columnHeaders,
-        data,
-        defaultCellActions,
-        disabledCellActions,
-        dispatch,
-        id,
-        pageSize,
-      ]
-    );
-
-    const onChangeItemsPerPage = useCallback(
-      (itemsChangedPerPage) => {
-        dispatch(tGridActions.updateItemsPerPage({ id, itemsPerPage: itemsChangedPerPage }));
-      },
-      [id, dispatch]
-    );
-
-    const onChangePage = useCallback(
-      (page) => {
-        loadPage(page);
-      },
-      [loadPage]
-    );
-
-    // Store context in state rather than creating object in provider value={} to prevent re-renders caused by a new object being created
-    const [activeStatefulEventContext] = useState({
-      timelineID: id,
-      tabType,
-      enableHostDetailsFlyout: true,
-      enableIpDetailsFlyout: true,
-    });
-
-    const useFetchAlertsData = () => {
-      // TimelineItem -> EcsFieldsResponse
-      // TODO: we're mangling the data into EcsFieldsResponse[] here, but when cells are rendered
-      // we're getting TimelineNonEcsData[] from the renderer. I don't think we're responsible
-      // for that unnecessary transformation. It might be something in triggers-actions-ui.
-      const alerts: EcsFieldsResponse[] = data.map((alert) => {
-        return alert.data.reduce(
-          (acc, curr) => {
-            acc[curr.field] = curr.value as string[];
-            return acc;
-          },
-          { _id: alert._id, _index: alert._index ? alert._index : '' } as EcsFieldsResponse
-        );
-      });
-
-      return {
-        activePage,
-        alerts,
-        alertsCount: totalItems,
-        // TODO: what does this control?
-        isInitializing: false,
-        isLoading,
-        // TODO: get actual data for this
-        getInspectQuery: () => ({ request: {}, response: {} }),
-        onColumnsChange: (columns: EuiDataGridColumn[], vc: string[]) => {
-          onSetVisibleColumns(vc);
-        },
-        // TODO: check the types here
-        onPageChange: (pagination: any) => onChangePage(pagination),
-
-        onSortChange: (s: EuiDataGridSorting['columns']) => onSort(s),
-        // TODO: check what refresh refers to
-        refresh: () => console.log('refresh'),
-        sort: sortingColumns,
-      } as FetchAlertData;
-    };
-
-    const config = triggersActionsUi.alertsTableConfigurationRegistry.get(
-      ALERT_TABLE_CONFIGURATION_KEY
-    );
-
-    const alertsTableProps = {
-      alertsTableConfiguration: config,
-      // TODO: why do we have to pass columns here again even though they're already part of the config?
-      columns: config.columns,
-      // defaultCellActions: TGridCellAction[];
-      deletedEventIds: [],
-      disabledCellActions: [],
-      // flyoutSize?: EuiFlyoutSize;
-      pageSize,
-      pageSizeOptions: itemsPerPageOptions,
-      // id?: string;
-      leadingControlColumns: [],
-      // TODO: render our flyout
-      showExpandToDetails: true,
-      trailingControlColumns: [],
-      useFetchAlertsData,
-      visibleColumns,
-      'data-test-subj': 'body-data-grid',
-      updatedAt: Date.now(),
-    };
-
-    return (
-      <>
-        <TGridComponentStateProvider
-          browserFields={browserFields}
-          columnHeaders={columnHeaders}
-          customBulkActions={additionalBulkActions}
-          filterStatus={filterStatus}
-          filterQuery={filterQuery}
-          indexName={indexNames.join()}
-          // TODO: this is rather hacky and we should clean this up for sure
-          // Accessing index 1 here means we are selecting the custom actions.
-          // Index 0 contains the checkbox column which we don't need with the
-          // new alert table.
-          customActionsColum={{
-            renderer: leadingTGridControlColumns[1].rowCellRender as (
-              props: EuiDataGridCellValueElementProps
-            ) => ReactNode,
-            width: leadingTGridControlColumns[1].width,
-          }}
-          timelineId={id}
-          renderCellValue={renderCellValue}
-          rowRenderers={rowRenderers}
-          showAlertStatusActions={showAlertStatusActions}
-        >
-          <StatefulEventContext.Provider value={activeStatefulEventContext}>
-            {
-              tableView === 'gridView' && <>{triggersActionsUi.getAlertsTable(alertsTableProps)}</>
-              // <EuiDataGridContainer hideLastPage={totalItems > ES_LIMIT_COUNT}>
-              //   <EuiDataGrid
-              //     id={'body-data-grid'}
-              //     data-test-subj="body-data-grid"
-              //     aria-label={i18n.TGRID_BODY_ARIA_LABEL}
-              //     columns={columnsWithCellActions}
-              //     columnVisibility={{ visibleColumns, setVisibleColumns: onSetVisibleColumns }}
-              //     gridStyle={gridStyle}
-              //     leadingControlColumns={leadingTGridControlColumns}
-              //     toolbarVisibility={toolbarVisibility}
-              //     rowCount={totalItems}
-              //     renderCellValue={renderTGridCellValue}
-              //     sorting={{ columns: sortingColumns, onSort }}
-              //     onColumnResize={onColumnResize}
-              //     pagination={{
-              //       pageIndex: activePage,
-              //       pageSize,
-              //       pageSizeOptions: itemsPerPageOptions,
-              //       onChangeItemsPerPage,
-              //       onChangePage,
-              //     }}
-              //     ref={dataGridRef}
-              //   />
-              // </EuiDataGridContainer>
-            }
-            {tableView === 'eventRenderedView' && (
-              <EventRenderedView
-                appId={appId}
-                alertToolbar={alertToolbar}
-                events={data}
-                leadingControlColumns={leadingTGridControlColumns ?? []}
-                onChangePage={onChangePage}
-                onChangeItemsPerPage={onChangeItemsPerPage}
-                pageIndex={activePage}
-                pageSize={pageSize}
-                pageSizeOptions={itemsPerPageOptions}
-                rowRenderers={rowRenderers}
-                timelineId={id}
-                totalItemCount={totalItems}
-              />
-            )}
-          </StatefulEventContext.Provider>
-        </TGridComponentStateProvider>
-      </>
-    );
+  ({ tableView = 'gridView', ...restProps }) => {
+    switch (tableView) {
+      case 'gridView':
+        return <StatefulGridBody {...restProps} />;
+      case 'eventRenderedView':
+        return <StatefuEventRenderedBody {...restProps} />;
+      default:
+        return null;
+    }
   }
 );
 
 BodyComponent.displayName = 'BodyComponent';
-
-const makeMapStateToProps = () => {
-  const memoizedColumnHeaders: (
-    headers: ColumnHeaderOptions[],
-    browserFields: BrowserFields
-  ) => ColumnHeaderOptions[] = memoizeOne(getColumnHeaders);
-
-  const getTGrid = tGridSelectors.getTGridByIdSelector();
-  const mapStateToProps = (
-    state: TimelineState,
-    { browserFields, id, hasAlertsCrud }: OwnProps
-  ) => {
-    const timeline: TGridModel = getTGrid(state, id);
-    const {
-      columns,
-      excludedRowRendererIds,
-      isSelectAllChecked,
-      loadingEventIds,
-      selectedEventIds,
-      showCheckboxes,
-      sort,
-      isLoading,
-    } = timeline;
-
-    return {
-      columnHeaders: memoizedColumnHeaders(columns, browserFields),
-      excludedRowRendererIds,
-      isSelectAllChecked,
-      loadingEventIds,
-      isLoading,
-      id,
-      selectedEventIds,
-      showCheckboxes: hasAlertsCrud === true && showCheckboxes,
-      sort,
-    };
-  };
-  return mapStateToProps;
-};
-
-const mapDispatchToProps = {
-  clearSelected: tGridActions.clearSelected,
-  setSelected: tGridActions.setSelected,
-};
-
-const connector = connect(makeMapStateToProps, mapDispatchToProps);
-
-type PropsFromRedux = ConnectedProps<typeof connector>;
-
-export const StatefulBody: React.FunctionComponent<OwnProps> = connector(BodyComponent);
